refactor: extract ui release notes helpers
